+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2026-05-02
+
+### Added
+
+- Added `/rremover folder` command to open the config folder.
+
+### Fixed
+
+- Fixed removed potions still being brewable.
+
 ## [3.0.1] - 2026-04-25
 
 ### Fix